Unify some error types that brought no value
