+++ rust/agama-server/src/error.rs
@@ -1,3 +1,10 @@
+use agama_lib::error::ServiceError;
+use axum::{
+    http::StatusCode,
+    response::{IntoResponse, Response},
+    Json,
+};
+use serde_json::json;
 use zbus_macros::DBusError;
 
 #[derive(DBusError, Debug)]
@@ -23,5 +30,22 @@ impl From<anyhow::Error> for Error {
 impl From<Error> for zbus::fdo::Error {
     fn from(value: Error) -> zbus::fdo::Error {
         zbus::fdo::Error::Failed(format!("Localization error: {value}"))
+    }
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum ApiError {
+    #[error("Agama service error: {0}")]
+    Service(#[from] ServiceError),
+    #[error("D-Bus error: {0}")]
+    DBus(#[from] zbus::Error),
+}
+
+impl IntoResponse for ApiError {
+    fn into_response(self) -> Response {
+        let body = json!({
+            "error": self.to_string()
+        });
+        (StatusCode::BAD_REQUEST, Json(body)).into_response()
     }
 }
