+++ rust/agama-server/src/software/web.rs
@@ -6,7 +6,7 @@
 //! * `software_stream` which offers an stream that emits the software events coming from D-Bus.
 
 use crate::{
-    error::Error,
+    error::{ApiError, Error},
     web::{
         common::{progress_router, service_status_router},
         Event,
@@ -22,15 +22,11 @@ use agama_lib::{
 };
 use axum::{
     extract::State,
-    http::StatusCode,
-    response::{IntoResponse, Response},
     routing::{get, post, put},
     Json, Router,
 };
 use serde::{Deserialize, Serialize};
-use serde_json::json;
 use std::{collections::HashMap, pin::Pin};
-use thiserror::Error;
 use tokio_stream::{Stream, StreamExt};
 
 #[derive(Clone)]
@@ -43,21 +39,6 @@ struct SoftwareState<'a> {
 pub struct SoftwareConfig {
     patterns: Option<Vec<String>>,
     product: Option<String>,
-}
-
-#[derive(Error, Debug)]
-pub enum SoftwareError {
-    #[error("Software service error: {0}")]
-    Error(#[from] ServiceError),
-}
-
-impl IntoResponse for SoftwareError {
-    fn into_response(self) -> Response {
-        let body = json!({
-            "error": self.to_string()
-        });
-        (StatusCode::BAD_REQUEST, Json(body)).into_response()
-    }
 }
 
 /// Returns an stream that emits software related events coming from D-Bus.
@@ -159,9 +140,7 @@ pub async fn software_service(dbus: zbus::Connection) -> Result<Router, ServiceE
     (status = 200, description = "List of known products", body = Vec<Product>),
     (status = 400, description = "The D-Bus service could not perform the action")
 ))]
-async fn products(
-    State(state): State<SoftwareState<'_>>,
-) -> Result<Json<Vec<Product>>, SoftwareError> {
+async fn products(State(state): State<SoftwareState<'_>>) -> Result<Json<Vec<Product>>, ApiError> {
     let products = state.product.products().await?;
     Ok(Json(products))
 }
@@ -185,7 +164,7 @@ pub struct PatternEntry {
 ))]
 async fn patterns(
     State(state): State<SoftwareState<'_>>,
-) -> Result<Json<Vec<PatternEntry>>, SoftwareError> {
+) -> Result<Json<Vec<PatternEntry>>, ApiError> {
     let patterns = state.software.patterns(true).await?;
     let selected = state.software.selected_patterns().await?;
     let items = patterns
@@ -216,7 +195,7 @@ async fn patterns(
 async fn set_config(
     State(state): State<SoftwareState<'_>>,
     Json(config): Json<SoftwareConfig>,
-) -> Result<(), SoftwareError> {
+) -> Result<(), ApiError> {
     if let Some(product) = config.product {
         state.product.select_product(&product).await?;
     }
@@ -237,7 +216,7 @@ async fn set_config(
 ))]
 async fn get_config(
     State(state): State<SoftwareState<'_>>,
-) -> Result<Json<SoftwareConfig>, SoftwareError> {
+) -> Result<Json<SoftwareConfig>, ApiError> {
     let product = state.product.product().await?;
     let product = if product.is_empty() {
         None
@@ -269,7 +248,7 @@ pub struct SoftwareProposal {
 ))]
 async fn proposal(
     State(state): State<SoftwareState<'_>>,
-) -> Result<Json<SoftwareProposal>, SoftwareError> {
+) -> Result<Json<SoftwareProposal>, ApiError> {
     let size = state.software.used_disk_space().await?;
     let proposal = SoftwareProposal { size };
     Ok(Json(proposal))
@@ -284,7 +263,7 @@ async fn proposal(
         (status = 400, description = "The D-Bus service could not perform the action
 ")
 ))]
-async fn probe(State(state): State<SoftwareState<'_>>) -> Result<Json<()>, SoftwareError> {
+async fn probe(State(state): State<SoftwareState<'_>>) -> Result<Json<()>, ApiError> {
     state.software.probe().await?;
     Ok(Json(()))
 }
