+++ rust/agama-server/src/web/common.rs
@@ -7,21 +7,13 @@ use agama_lib::{
     progress::Progress,
     proxies::{ProgressProxy, ServiceStatusProxy},
 };
-use axum::{
-    extract::State,
-    http::StatusCode,
-    response::{IntoResponse, Response},
-    routing::get,
-    Json, Router,
-};
+use axum::{extract::State, routing::get, Json, Router};
 use pin_project::pin_project;
 use serde::Serialize;
-use serde_json::json;
-use thiserror::Error;
 use tokio_stream::{Stream, StreamExt};
 use zbus::PropertyStream;
 
-use crate::error::Error;
+use crate::error::{ApiError, Error};
 
 use super::Event;
 
@@ -81,21 +73,6 @@ struct ServiceStatusState<'a> {
 struct ServiceStatus {
     /// Current service status.
     current: u32,
-}
-
-#[derive(Error, Debug)]
-pub enum ServiceStatusError {
-    #[error("Service status error: {0}")]
-    Error(#[from] ServiceError),
-}
-
-impl IntoResponse for ServiceStatusError {
-    fn into_response(self) -> Response {
-        let body = json!({
-            "error": self.to_string()
-        });
-        (StatusCode::BAD_REQUEST, Json(body)).into_response()
-    }
 }
 
 /// Builds a stream of the changes in the the `org.opensuse.Agama1.ServiceStatus`
@@ -186,24 +163,7 @@ struct ProgressState<'a> {
     proxy: ProgressProxy<'a>,
 }
 
-#[derive(Error, Debug)]
-pub enum ProgressError {
-    #[error("Progress error: {0}")]
-    Error(#[from] ServiceError),
-    #[error("D-Bus error: {0}")]
-    DBusError(#[from] zbus::Error),
-}
-
-impl IntoResponse for ProgressError {
-    fn into_response(self) -> Response {
-        let body = json!({
-            "error": self.to_string()
-        });
-        (StatusCode::BAD_REQUEST, Json(body)).into_response()
-    }
-}
-
-async fn progress(State(state): State<ProgressState<'_>>) -> Result<Json<Progress>, ProgressError> {
+async fn progress(State(state): State<ProgressState<'_>>) -> Result<Json<Progress>, ApiError> {
     let proxy = state.proxy;
     let progress = Progress::from_proxy(&proxy).await?;
     Ok(Json(progress))
